+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * `getTemplateById` - get the latest version of a template by id.
 * `getTemplateVersion` - get the template by id and version.
 * `getAllTemplates` - get all templates, can be filtered by template type.
-* `getTemplatePreview` - get the contents of a template with the placeholders replaced with the given personalisation.
+* `generateTemplatePreview` - get the contents of a template with the placeholders replaced with the given personalisation.
 * See the README for more information about the new template methods.